CRM-11925 - CRM_Utils_Hook - Remove moduleImplements.
The Drupal implementation is good, but we don't need the functionality, and
it's a bit of a liability because the design isn't portable to Joomla.
Keeping it might encourage others to write unportable hook code.

However, if we find that we do need the functionality, then we should use
this code as a starting point.  (Aside: For a non-Drupal example, the
WordPress implementation is better than the Joomla implementation.  The WP
version looks like it would still work with Civi ext-modules.)