CRM-12146 - Add test case for Group.get(group_id=>1) and Contact.get(contact_id=>1).
I tried inspecting the DAO_*::fields() metadata to determine which entities
and field names to test.  This can be done (if you see that the array-keys
are unique-names and the 'name' properties ar enon-unique-names), but it
doesn't ultimately help because the XML schema for group doesn't officially
list group_id as a unique-name.  I'm nervous that changing that schema would
have unintended consequences, so...  this patch uses a more conservative
approach (whitelisting/hardcoding certain entities/unique-names).