De-hack additional operators being added for report class only
The buildDate range function is called about 20 times with no additional
operators being added & twice with them being added. But the default
behaviour was to add them. This makes the default behaviour NOT to add
them & makes the one class that does add them take responsibility for
passing them in rather than a round about method of deriving them