CRM-12943 - Move Collection.fetchCreate() to CRM.Backbone.findCreate()
This may be a subtle point, but in the current use-cases, we don't really
have a "collection" that we're trying to maintain.  We're just using the
collection to facilitate the search.  If have this helper a method of the
collection, then there are design questions about how one uses the
collection (eg with repeated calls to fetchCreate), but that's all
extraneous.

----------------------------------------
* CRM-12943: Make HTML prototype of job UI functional
  http://issues.civicrm.org/jira/browse/CRM-12943