CRM-13103 - CRM_Utils_HookTest
Add unit test in which one hook calls another. Note that this unit-test
doesn't actually reproduce the issue for me (eg it incorrectly shows the
test as passing on code that doesn't work for me in real situations), but
it's still valuable in that it gives us better test coverage.

----------------------------------------
* CRM-13103: Hook execution terminates prematurely
  http://issues.civicrm.org/jira/browse/CRM-13103