Make `formal_title` field optional
Add a config setting to make display of the Title input field (in
contact edit forms) optional.

(This doesn't affect contact view, as the name displayed there is
determined by the "Individual Display Name Format" template instead.)

This is a minimal approach, adding just the title setting. Perhaps it
might be preferable to add a full set of options for disabling the other
name fields as well, while at it?...

(We are willing to do that too, as it shouldn't be too hard, and we have
users for that as well -- but we need some consensus about the right way
for handling this.)

The setting is placed in the "Display Preferences", next to the
"Individual Display Name Format". However, it could be placed next to
"Editing Contacts" (also in "Display Preferences") as well -- it's
somewhat related to both...