Fix localisation error in DB upgrade script
The upgrade script for 4.3.6 has a hard-coded lookup for the 'Credit
Card' payment_instrument, but uses the (default) `label` column for the
lookup, which is localised -- i.e. the upgrade fails fatally on any
setup where the 'Credit Card' label is translated.

Fix by using the non-localised `name` column instead.