Easy to run tests.
Made a new script (./tests/run.php) that when run out the civicrm-core
repository can setup most eveything needed to run tests and then run the
tests. For the unit tests, it should just run with no dependencies, but for
the Web (Selenium) tests, there are the following requirements:
drush
firefox
apache (or some other HTTP server)

With all that available, it will download and setup Drupal and suggest Apache
settings.

This was developed on Ubuntu, but I would like to get it working on other
platforms.

Run it with -h to see the options.