CRM-14221 - Improving Payment Express/DPS payment processor strict standards compatibility
I'm running a fairly recent version of PHP, and have strict standards turned on.. so the Payment Express component was spewing warnings and errors about non-static functions being called statically and the singleton function not being compatible with its parent.

The solution, make the CRM_Core_Payment_PaymentExpressUtils functions static, and change the definition of the CRM_Core_Payment_PaymentExpress and CRM_Core_Payment_PaymentExpressIPN singleton() functions to match that in CRM_Core_Payment

I also discovered an undefined CURL option in CRM_Core_Payment_PaymentExpressUtils that was being set - This doesn't do anything and causes a warning so I removed it.