CRM-14268 - Make CiviCRM's event registration cart record financial transactions to conform to the changes beginning in CiviCRM 4.2.
The newer versions of Civi create a civicrm_line_item for every participant
registered for an event.  This was happening automatically for cart
registrations, but the records created were not correct as Civi was
autogenerating them with default values.  This commit corrects that and
creates accurate civicrm_line_item records for each participant
registered in the cart.

The newer versions of Civi also automatically create records in
civicrm_entity_financial_trxn for each civicrm_contribution and civicrm_financial_item.  The cart code manually generated a civicrm_entity_financial_trxn for each contribution, which was causing some extra records to be created with incorrect values. Removing the manual generation of civicrm_entity_financial_trxn records in the cart code fixes this issue.  The correct rows now appear in civicrm_entity_financial_trxn.

We also want to prevent the cart from creating an extra civicrm_financial_trxn record. These records are now auto generated and we no longer have to manually create a record.  Manually doing this caused an extra record to be created in civicrm_financial_trxn set to the total of all the contributions from the cart, which is incorrect.