Reimplemented CodeGen to work from Doctrine metadata.
We are going to move over to Doctrine and as part of that we want to
have the database defined in only one place. Before Doctrine we were
using a proprietary XML format to store the database schema. Now we
are going to use annotations in the Doctrine entities to define the
schema.

But, we aren't going to cut over to Doctrine all in one go, so we need
to still generate the DAO files which were generated from the XML
format. This changes the code generator to generate the DAOs from the
Doctrine entity metadata instead of the xml files.

Now uses 'bin/gencode' to run GenCode.

Some of the metadata from the XML files isn't coming over in this
version like comments, unique_name, rows, etc... We're not sure what
we're going to do about that yet.