Can't use file_exists to check for classes.
Classes could be in any of the paths. When running the tests, some of
the CRM_* classes are in tests/phpunit/CRM and this prevents them from
loading. If we want to do this, then we need to check all the include
paths, but require_once does that for us, so I'm not sure why we
wouldn't just want to let require_once fail if it can't find the file.