Enity and DAO generation fixes.
I removed the skipTables part of entity generation because it gets into
some kind of infinite loop and I don't think we need to skip it. I fixed
the problem with Case entities, and we can probably deal with the
entity_table problem as well.

Generated entities need to put \ before any CRM_ classes or it can't
find them in the namespace.

Need the $_fields variable to be declared or we get an error.

Changed the listAll template back to use the DAO entity data because
it's easier.