Fixed a problem with DAO classes not existing.
This check was dying because DAO classes might not exist yet they don't
get loaded by the class loader until some code tries to use them. I
think this was working before because the old DAO classes required all
their related foreign key classes, but I took that out to rely on the
autoloader.

I think this is probably better anyway, because we probably want an
exception in most cases where the class doesn't exist and we don't want
to fail silently.