Fix the retrieveContact test fail on asserting relatiotionship for current employer.
The test was checking for a relationship to an organization contact with an
ID of 1, but that was the ID of the individual contact record that's created
at the start of this test function. I've fixed this by explicitly creating an
organization contact and using that for the current employer relationship, so
we always know what the contact ID of the organization record will be.