distmaker - Allow running individual builds without switching branches
One of the more annoying things with distmaker is that it switches all the
branches, which means that you need to commit everything before building.
This is necessary for building D6 & D7 in the same script, but makes it
harder to work on changes to builds.

With this patch, one can directly call the "dists/*.sh" files to trigger a
build, and it muck with git.  All the git checkouts are now specified in the
distmaker.sh