CRM-14856 - CRM_Profile_Form - Restore behavior of add_to_group_id when there are multiple profiles
The behavior is to scan a combined list of UFFields, see if each UFField's
parent UFGroup has "add_to_group", and use it if found.  If multiple
UFGroups define "add_to_group", then the value of the last UFField wins.

I don't see this as good behavior -- just compatible behavior.