Test for SA-2014-002.
This is just the test for SA-2014-002 because I want the fix in a
separate commit that I can apply without the tests because I had to make
some fairly low level changes to CRM_Utils_System to get the test to work,
and I don't necessarily want to apply that patch to 4.3 which is more a
maintenance release.

If an anonymous user created a request like this:
http://civicrm.local/civicrm/ajax/permlocation?cid=1&ufId=2

It would return some JSON containing contact information for the contact
specified by the cid parameter. This would let anyone get that information
for any contact. Only users who have been granted permission to view that
information should be able to do this.

Added a test for this problem. Also I changed a few static methods in
CRM_Utils_System to call methods on $config->userSystem. This lets us
override those methods when doing UnitTests. We need to do this for
functions like CRM_Core_Page_AJAX_Location::getPermissionedLocation
because it calls CRM_Utils_System::civiExit which calls the PHP
function exit which then exits phpunit.

Needed similar things for CRM_Utils_System::redirect and
CRM_Utils_System::bounce which mess with the Apache headers and call PHP
exit which both cause problems with phpunit tests.