Fix for SA-2014-002.
If an anonymous user created a request like this:
http://civicrm.local/civicrm/ajax/permlocation?cid=1&ufId=2

It would return some JSON containing contact information for the contact
specified by the cid parameter. This would let anyone get that information
for any contact. Only users who have been granted permission to view that
information should be able to do this.