CRM-14856 - A test for this issue.
A test for the case where the user changes the cancelURL field to a
different URL than specified in the settings for a profile. We do not want
to allow this.

A test for the case where the user changes the add_to_groups field to be a
gid for a different group than the one allowed in the profile.

A few changes to the Utils_System classes to help with writing unit tests
for internal methods.