Revert "CRM-14946 add relevant test data set"
After interviewing Eileen, Kurund, and Dave, it appears that 450a63c was a
draft that shouldn't have been merged.  The underlying question was how to
handle a change in the smart-group data-structure.  One approach (implicit
in 450a63c) was to maintain runtime compatibility.  Another approach (which
was ultimately chosen) was to define an upgrade that altered the
data-structure.

This fixes the failure in CRM_Contact_BAO_GroupTest::testGroupData.

This reverts commit 450a63cfc27064f2ffc141e76d4662fe612a67a5.