Now localization info is transferred correctly to DAO objects.
We were missing the localization information in the new stuff because it
doesn't get passed in the fields() array. I added it as a new annotation
in the Entities called Field (e.g. Field(localization=true)).

I also ran some tests to check that the new DAO objects generate exactly
the same ::fields() arrays as they used to.