Add a new findById DAO method.
We have found it very useful to have this simple function, which retrieves a
DAO object of a specific type by ID or throws an exception if one is not
found.

This can be called by using something like
CRM_Contribute_BAO_Contribution::findById(35), which will retrieve a
Contribution Object for that id, if it exists.