CRM-15578 - Pull up static $mailsProcessed and reset during testing
These static variables were leaking between test-cases in a way which caused
the second test to get stuck on an infinite loop.  It has been quite
difficult to identify the tests which are interacting
(api_v3_JobProcessMailingTest and api_v3_MailingTest), the particular loop
which went infinitely, and the leaky variables.

I don't understand why these statics exist, but I don't have time to fully
grok it -- right now, the branch "master-civimail-abtest" has gone off on a
bit of a limb with failing tests, I just want to bring it back to a stable
place.  Promoting & resetting the static is the simplest way.