CRM-15578 - crmMailing(AB) - Submission & status tracking
When loading or submitting, determine the status of any jobs tied to this mailing.

When attempting to edit a submitted mailing, display a message instead of a form.

Change the technique for updating URL -- instead of watching 'mailing.id',
explicitly call updateUrl after saving.  This ensures a consistent, linear
sequence of operations.  (Note: The watch technique was useful when there
were different variations of save() in EditMailingCtrl and
PreviewMailingCtrl, but we eliminated PreviewMailingCtrl, so it's much
easier to manage the dataflows in a centralized fashion.)