crmMailingAB - Transplant report. Decouple stats API.
Note: Previously, the MailingAB.graph_stats API keyed off the pre-declared
"winning criteria" and target date, which made it necessary to choose these
things in advance.  This API now accepts these as parameters which can be
tweaked interactively (based on what the user whats to investigate).

The graphs don't really work right now, but I've transplanted the code to
show how to encapsulate it in a directive.