Get the distmaker working with npm and bower.
We're going to start distributing the node_modules and bower_modules
using npm and bower now instead of having them in packages because we
feel that those are a better way to manage backages than the packages
repository. So the distmaker needs to copy those directories to the
distribution after running 'npm install' to get the packages downloaded.

Also I needed to put a resolution for the angular version in the
bower.json file because some of the angular plugins were resolving to
angular 1.2.x and we want 1.3.x. I also changed the version numbers to
start at 1.3.8, but allow minor version updates because that's what
we're using.