Fixed drush and node_modules/*.info file conflict
Tim noticed that drush doesn't work when node_modules directory exists.
People say that it's because of the *.info files that some things have
in node_modules and suggested adding a line to the npm postinstall
script to remove  any *.info files from there because they aren't really
needed to run stuff. Because we are already doing brower install in the
postinstall script, I broke the postinstall stuff into a separate
script and put both things in there.