CRM-15832 - CRM_Core_Resources - Move translateScript to class. Add HTML support.
For the moment, it looks like the existing JS string-extractor actually
works for Angular-style HTML partials.  This is not entirely a coincidence
(since our Angular convention was purposefully derived from our JS
convention), but luck/circumstance is a big factor, and this may not work in
the long-run.  This commit makes it easy to use different string-extractors
for different file-types.