CRM-15832 - Fix domain used by core Angular apps
The old code was based on a misunderstanding. I thought that we had separate
translation domains for each Civi component.  This is half true -- when the
translators work on them, there are separate .po files for each component.
However, when we distribute the final .mo files, the components are
aggregated together into one "civicrm.mo".  Thus, at runtime, there is
really only one domain for the core app.

There are other domains the extensions, and there are a couple contexts
within the civicrm domain (eg for states/provinces), but that's irrelevant
to core's CiviCase and CiviMail UIs.