CRM-15931, CRM-15733 - civicrm_mailing_group.group_type - Restore uppercase values
In 4.5.x, the schema for group_type changed from
enum('Include','Exclude','Base') to varchar, which made it more sensitive to
variations in case.  New mailings would be created with lowercase 'include',
but old mailings still had 'Include'.  UI's would display correctly for old
mailings but not for new mailings.

CRM-15733 and CRM-15931 both report peculiar symptoms of the change.

This patch aims to consistently use the 4.4.x data-model (for both old and
new mailings).