CRM-15856 - crmMailing - Display a warning inline if required tokens are missing
When you first fill in the body, we produce a popup on-blur to mention the
missing tokens.  However, when you edit an existing mailing, there is no
blur at first, and it's not clear what's wrong (unless you focus and blur
the problematic field).  This makes the problem more obvious.