Make "Chain Select" mechanism robust against missing 'target' form field
If removeElement() was invoked on an element that is the 'target' of a
'chainSelect' (which can happen through a form hook in an extension),
the chainSelect handling code was subsequently erroring out.

Add checks to avoid that.

This is related to the CRM-15476 fixes in
8f9c3cbea992320f344d576c8a87c5bab318466c and
14b2ff1521a8c25e363b13723c2979f52ad1cd47 , which already fixed the
situation where the 'control' element is missing.