CRM-16120 - crmUiDebug - Only display debug fields if ?angularDebug=1
We've been outputting a collapsible debug panel by default. This completely
hides (or shows) the debug panel depending on whether parameter
"angularDebug" is set.

Note that there are also browser plugins which are richer, more flexible,
more thorough, and it's probably better to install one of those.  However,
it's tough to standardize on one plugin (b/c they're different for each
browser and because we can't bundle them into buildkit).  In a crunch (e.g.
when debugging on a non-developer's workstation), the parameter could still
be helpful.