Fix conference slot does not seem to save parent item. CRM-16100
The conference slot parent event field was not saving the parent event data.
The new version of Civi automatically looks up the event information from the
event id saved in that field, so I removed a bunch of old logic that did all
that for us. The only thing that needs to be saved in that field now is the
parent event id, which is what is happening.