CRM-16145 - When aggregating JS, dedupe adjacent closures.
The reorganized Angular files have added ~50 identical closures, and there
are still other big modules which remain to split.  Ideally, the raw JS
files wouldn't bother with these closures, but we'd have to rework the debug
use-case and karma use-case if we removed the closures from the original
file (and that most likely means a change to developer workflow).  This
patch accomplishes the goal without changing developer workflow.