The text string is very difficult to translate
This string should either be rephrased or have a comment attached, so Transifex translators have a change of understanding what is meant here. I don't know what to make of it translating through Transifex. In Transifex the translators don't get access to the comments, so if possible, please phrase this string in a very explicit way.