CRM-17003 previous commit caused wrong formatting in access tab on report
The issue appeared to be that the re-formatting of administer multisite is happening after the re-formatting
dependent on the  parameter being passed in. The function doing the reformatting is apparently only called from
that function so we could move the description processing to the calling function.

NB - need to grep wordpress, joomla, drupal dirs for that fn before making the change I guess,
Or we could just do the  stuff twice to be safe