CRM-13725 - Extra (failing) unit tests.
If you have an existing relationship and a new relationship,
and the new relationship has the same standard fields than the existing one,
but only one of the two relationships has custom fields, CiviCRM
will not accept the new relationship, because it will consider the
new one as a duplicate.

Two unit tests, they will obviously fail.

----------------------------------------
* CRM-13725: Make check for duplicate relationship aware of custom fields
  https://issues.civicrm.org/jira/browse/CRM-13725