CRM-17191 - ensure Participant list report works with custom Grouping
Previously, we got a trace back if a user defined a group by clause.
I've removed what appears to be crufty code that tries to add the
group by fields to an ORDER BY statement - which conflicts with the
existing ORDER BY statement and seems unnecessary. Also, if the user
wants to group by another field, we should not group by the participant
id.