CRM_Utils_Cache::create - Use factory function
Low-level tools like GenCode may run without access to a database. When we
try to use caching in central code, the cache-accesses are prone to
crashing. The problem -- in this environment, we shouldn't even try to use
databse for caching.

`CRM_Utils_Cache::create()` will attempt to create the best cache it can,
depending on what services are available.